Refactor com.viwilo to org.apache.camel
